+++ app/src/components/pages/searchPage/searchPage.tsx
@@ -1,12 +1,9 @@
 "use client";
 import {
   Box,
-  Text,
   Pagination,
   Heading,
   Flex,
-  HorizontalRule,
-  TagSet,
   Link,
   Icon,
   Menu,
@@ -17,6 +14,7 @@ import { displayResults, totalNumPages } from "@/src/utils/utils";
 import Filters from "../../search/filters/filters";
 import { useSearchContext } from "@/src/context/SearchProvider";
 import { usePathname, useRouter } from "next/navigation";
+import ActiveFilters from "../../search/filters/activeFilters";
 import SearchCardsGrid from "../../grids/searchCardsGrid";
 import { headerBreakpoints } from "@/src/utils/breakpoints";
 
@@ -78,22 +76,7 @@ const SearchPage = ({ data }) => {
           },
         }}
       >
-        <Flex alignContent="center" gap="xs">
-          <Text size="subtitle2" sx={{ margin: 0, fontWeight: 400 }}>
-            Filters applied:
-          </Text>
-          <TagSet
-            isDismissible
-            id="search-filter-tags"
-            onClick={() => {}}
-            tagSetData={[
-              { id: "audio", label: "Audio" },
-              { id: "video", label: "Video" },
-            ]}
-            type="filter"
-          />
-        </Flex>
-        <HorizontalRule />
+        <ActiveFilters />
         <Flex
           sx={{
             [`@media screen and (min-width: ${headerBreakpoints.lgMobile}px)`]:
@@ -112,6 +95,7 @@ const SearchPage = ({ data }) => {
         >
           <Heading
             size="heading5"
+            ref={headingRef}
             tabIndex={-1}
             margin="0"
           >{`Displaying ${displayResults(