+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added custom select component, `selectFilter`, and corresponding modal for facet filters (DR-3394, DR-3396)
+- Added search result model, mock data, card component, and card grid component (DR-3363)
+
 ### Updated
 
 - Updated collection search to use `q` param and added redirect for `collection_keywords` (DR-3337)